+++ c3smembership/administration.py
@@ -24,6 +24,7 @@ from c3smembership.models import (
     Group,
 )
 from c3smembership.presentation.views.dashboard import get_dashboard_redirect
+from c3smembership.views.afm import send_mail_confirmation_mail
 
 
 DEBUG = False
@@ -216,85 +217,19 @@ def mail_mail_conf(request):
     Needed for applications that came in solely on paper
     and were digitalized/entered into DB by staff.
     '''
-    afmid = request.matchdict['member_id']
-    afm = C3sMember.get_by_id(afmid)
-    if isinstance(afm, NoneType):
+    member_id = request.matchdict['member_id']
+    member = C3sMember.get_by_id(member_id)
+    if isinstance(member, NoneType):
         request.session.flash(
             'id not found. no mail sent.',
             'messages')
         return get_dashboard_redirect(request)
 
-    import random
-    import string
-    _looong_token = u''.join(
-        random.choice(
-            string.ascii_uppercase + string.digits) for x in range(13))
-    _url = (request.registry.settings['c3smembership.url'] +
-            '/vae/' + afm.email_confirm_code +
-            '/' + _looong_token + '/' + afm.email)
+    send_mail_confirmation_mail(
+        member,
+        request.registry.settings['c3smembership.url'],
+        request.registry.get_mailer(request),
+        request.localizer,
+        request.registry.settings['testing.mail_to_console'])
 
-    _body = u'''[english version below]
-
-Hallo {1} {2},
-
-gemeinsam mit Dir freuen wir uns auf die erste Generalversammlung der C3S SCE
-am 23. August, um 14:00 Uhr im Millowitsch-Theater in Köln. Details dazu
-erhältst Du in Kürze in einer separaten Einladung.
-
-Da wir die Einladungen per E-Mail verschicken werden, möchten wir Dich bitten
-uns kurz zu bestätigen, dass diese E-Mail-Adresse korrekt ist und Du auf diesem
-Wege erreichbar bist. Dafür brauchst Du nur den folgenden Link aufzurufen:
-
-  {0}
-
-Solltest Du nicht {1} {2} sein und diese Adresse nicht bei uns angegeben haben,
-antworte bitte kurz auf diese E-Mail. Dann werden wir die Adresse aus unser
-Datenbank streichen.
-
-Antworte bitte ebenfalls, falls Du die E-Mail-Adresse ändern möchtest.
-
-
-Viele Grüße :: Das C3S-Team
-
-++++++++++++++++++++++++++++++++++++++++++++++++++
-
-Hello {1} {2},
-
-together with you we are happily awaiting the first general assembly of C3S SCE
-on August 23rd, at 2 pm in the Millowitsch-Theater in Cologne. You will soon
-receive the details in a separate invitation.
-
-As we will send the invitations via email, we would like you to confirm your
-email address beforehand.
-
-Just click the following link to confirm your address:
-
-    {0}
-
-If you are not {1} {2} and you didn't give this email address to us, please
-reply to this email with a short explanation. Then we will remove your address
-from our database.
-
-Should you want to change your email address please reply to this mail, too.
-
-Best wishes :: The C3S Team
-    '''.format(
-        _url,  # {0}
-        afm.firstname,  # {1}
-        afm.lastname,  # {2}
-    )
-
-    LOG.info("mailing mail confirmation to AFM # %s", afm.id)
-
-    message = Message(
-        subject=(u'[C3S] Please confirm your email address! '
-                 u'/ Bitte E-Mail-Adresse bestätigen!'),
-        sender='yes@c3s.cc',
-        recipients=[afm.email],
-        body=_body
-    )
-    mailer = get_mailer(request)
-    mailer.send(message)
-    afm.email_confirm_token = _looong_token
-    afm.email_confirm_mail_date = datetime.now()
-    return get_dashboard_redirect(request, afm.id)
+    return get_dashboard_redirect(request, member.id)